% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{AutoCorrLength}
\alias{AutoCorrLength}
\title{Calculate the Autocorrelation Length}
\usage{
AutoCorrLength(Y, nlag)
}
\arguments{
\item{Y}{Numeric vector. The time series for which the autocorrelation length is calculated.}

\item{nlag}{Integer. The maximum number of lags to consider in the autocorrelation function.}
}
\value{
A single numeric value representing the autocorrelation length.
}
\description{
This function calculates the autocorrelation length, which measures the
degree of autocorrelation in a time series. It uses the squared values
of the autocorrelation function (ACF) up to a specified number of lags.
The implementation is adapted from Straatsma et al. (2016).
}
\examples{
Y <- rnorm(100)  # Example time series
nlag <- 50       # Number of lags
autocorr_length <- AutoCorrLength(Y, nlag)
print(autocorr_length)

}
\references{
Straatsma, T. P., Berendsen, H. J. C., & Stam, A. J. (2016).
Estimation of statistical errors in molecular simulation calculations.
http://doi.org/10.1080/00268978600100071

Afyouni, Soroosh, Stephen M. Smith, and Thomas E. Nichols.
"Effective Degrees of Freedom of the Pearson's Correlation Coefficient
under Serial Correlation." bioRxiv (2018): 453795.
}
