% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nearestSPD}
\alias{nearestSPD}
\title{Convert a Matrix to the Nearest Positive Semi-Definite Matrix}
\usage{
nearestSPD(A)
}
\arguments{
\item{A}{Numeric matrix. A square matrix to be converted to its nearest PSD matrix.}
}
\value{
A numeric matrix that is positive semi-definite and as close as possible to the input matrix \code{A}.
}
\description{
This function converts a square matrix to the nearest positive semi-definite (PSD) matrix
using the algorithm described by Higham (1988). The input matrix is symmetrized and adjusted
iteratively to ensure all eigenvalues are non-negative.
}
\examples{
# Example of converting a non-PSD matrix to PSD
mat <- matrix(c(4, -2, 2, -2, 3, -1, 2, -1, 2), nrow = 3, byrow = TRUE)
psd_mat <- nearestSPD(mat)
eigen(psd_mat)$values  # All eigenvalues should now be non-negative

}
\references{
Higham, N. J. (1988). Computing the nearest symmetric positive semi-definite matrix.
Linear Algebra and its Applications, 103, 103-118.
}
