% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sim_ar1}
\alias{sim_ar1}
\title{Simulate a Single AR(1) Time Series}
\usage{
sim_ar1(Tt, phi)
}
\arguments{
\item{Tt}{Integer. Number of time points in the time series.}

\item{phi}{Numeric. AR(1) coefficient (typically between -1 and 1 for stationarity).}
}
\value{
A numeric vector of length \code{Tt} representing the simulated AR(1) time series.
}
\description{
Generates a single autoregressive time series of order 1 (AR(1)) with
specified length and autoregressive coefficient. The series is initialized
with a standard normal value and evolves as:
\deqn{Y_t = \phi Y_{t-1} + \epsilon_t, \quad \epsilon_t \sim \mathcal{N}(0, 1)}
}
\examples{
sim_ar1(Tt = 100, phi = 0.8)

}
\seealso{
\code{\link{sim_many_ts_same_ar1}} for simulating multiple AR(1) series.
}
